Readme added. Virtual play and simulation separated.
