+++ .gitignore
@@ -1,2 +1,5 @@
 client_secret.json
 logs/*.txt
+dist/*
+build/*
+
